Added balance checking to getWagerAmount

+++ src/main/java/com/github/zipcodewilmington/casino/GamblingGameInterface.java
@@ -2,5 +2,5 @@ package com.github.zipcodewilmington.casino;
 
 public interface GamblingGameInterface extends GameInterface{
 
-    int getWagerAmount();
+    double getWagerAmount();
 }
